refactor(forget-password-reset-password): organizing components in folders
