+++ src/app/[locale]/auth/forget-password/_components/OTP/otp.tsx
