+++ src/app/[locale]/auth/forget-password/_components/forget-password/forget-password.tsx
